Refactor: split goal input logic to custom hook
