+++ src/components/Sidebar/GoalList/index.tsx
@@ -1,43 +1,13 @@
-import { ChangeEvent, useEffect, useRef } from 'react';
-
 import { Input } from '@/components/common/Input';
 import { GoalItem } from '@/components/Sidebar/GoalList/GoalItem';
 
+import { useGoalInput } from '@/hooks/useGoalInput';
 import { useGoalsStore } from '@/store/useGoalStore';
 
 export const GoalList = () => {
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const { goals, setNewGoal, isNew, toggleIsNew, addGoal, newGoal } =
-    useGoalsStore();
-
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setNewGoal(e.target.value);
-  };
-
-  const handlePostNewGoal = () => {
-    addGoal();
-    setNewGoal('');
-    toggleIsNew(false);
-  };
-
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter' && newGoal.trim()) {
-      handlePostNewGoal();
-    }
-  };
-
-  const handleBlur = () => {
-    if (newGoal.trim()) {
-      handlePostNewGoal();
-    }
-  };
-
-  useEffect(() => {
-    if (inputRef.current) {
-      inputRef.current.focus();
-    }
-  }, [isNew]);
+  const { goals, isNew, newGoal } = useGoalsStore();
+  const { inputRef, handleInputChange, handleKeyDown, handleBlur } =
+    useGoalInput();
 
   return (
     <div className="flex w-full flex-col px-8">
